+++ src/client/index.js
@@ -3,10 +3,14 @@ import DomElement from "./element/DomElement";
 import Binder from "./binder";
 import FormValidator from "./form-validator";
 import EVUnit from "./form-validator/EVUnit";
+import Carousel from "./bootstrap-carousel";
+import CarouselItem from "./bootstrap-carousel/CarouselItem";
 export {
   PopupWindow,
   DomElement,
   Binder,
   FormValidator,
-  EVUnit
+  EVUnit,
+  Carousel,
+  CarouselItem
 };