+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 9/27/2019
+#### Unreleased
+
+### Creates
+
+- Add `Carousel` and `CarouselItem`
+
+### Updates
+
+- Update `DomElement` to make `append` and `prepend` able to deal with Both Html-Dom-Element or class `DomElement`
+
 # 9/20/2019
 
 #### [1.2.2](https://github.com/1846689910/cicero/commit/d60af15824a35d027f01317df9426da99be4c19e)